ingest/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
